+++ packages/liteforge/src/client.ts
@@ -1,9 +1,0 @@
-export * from '@liteforge/client';
-
-import type { Client } from '@liteforge/client';
-
-declare module '@liteforge/runtime' {
-  interface PluginRegistry {
-    client: Client;
-  }
-}